Create a generic dsi api response to spreadsheet class
add_dsi_approvers_to_spreadsheet and add_dsi_users_to_spreadsheet classes are almost identical, extracted a base class and implemented it in these usecases to remove duplication.
